Refactor ContactCard component
